requirements.txt
gradio>=4.19.0
openai>=1.12.0
langchain>=0.1.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
pdfplumber>=0.10.3
unstructured>=0.12.0
langchain-community>=0.0.20
langchain-openai>=0.0.8
langchain-text-splitters>=0.0.1
sentence-transformers>=2.2.2
huggingface-hub>=0.20.0
tiktoken>=0.5.2
numpy>=1.26.0
pypdf>=3.17.0